Add useCertificateDeleteDialog
