+++ src/dialogs/certificate-delete-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from "./useCertificateDeleteDialog";
